task: moved UserFilter and FilterButton to shell/components
