+++ src/shell/components/Filters/index.ts
@@ -1,0 +1,4 @@
+import { FilterButton } from "./FilterButton";
+import { UserFilter } from "./UserFilter";
+
+export { FilterButton, UserFilter };
